debugging author and user ( repo/service)

+++ app/repo/model.go
@@ -8,9 +8,9 @@ import (
 
 type BaseModel struct {
 	CreatedAt time.Time      `gorm:"column:created_at;autoCreateTime"`
-	CreatedBy int64          `gorm:"column:created_by"`
+	CreatedBy int            `gorm:"column:created_by"`
 	UpdatedAt time.Time      `gorm:"column:updated_at;autoUpdateTime"`
-	UpdatedBy *int64         `gorm:"column:updated_by"`
+	UpdatedBy *int           `gorm:"column:updated_by"`
 	DeletedAt gorm.DeletedAt `gorm:"column:deleted_at"`
-	DeletedBy *int64         `gorm:"column:deleted_by"`
+	DeletedBy *int           `gorm:"column:deleted_by"`
 }
